Create new function to show cart of a user.
Create a new cart.jsp file in users folder to show the cart of user.
Create a new CartServlet to handle request from user for access and redict.
Create new method in CategoryDAO to get a list of product which split into each category.
Create new CartDAO to get data which relative with cart from database.