Add fuction to show image from storage
In HomeServlet, added fuction to check out if the url is exist on the database or not. If not then add url to the database.
Fixed some query to get image path in ProductDAO.
Added fuction to show image from storage in index.jsp.
Create ImagePAathInitializer.java to handle setting url of product img and category img in database.