Add initialize for user image
Created method to get all id user from database and method to set user image path in database in UserDAO.
Created User constructor only need user id in User model.
Created method to check exists user image folder and create it if not exist and method to set image of user if exists in database when run the web.